removed some boxed primitives

+++ src/main/java/com/igknighters/util/Tracer.java
@@ -128,7 +128,7 @@ public class Tracer {
                 entry = entryHeap.get(trace.getKey());
             }
             entry.setDouble(trace.getValue());
-            Logger.recordOutput("Tracer/" + trace.getKey(), trace.getValue());
+            Logger.recordOutput("Tracer/" + trace.getKey(), trace.getValue().doubleValue());
         }
         traceTimes.clear();
     }
